Return error if API token required but not found
Currently, if you run a command without specifying a token (in config, as env var or arg), it appears as if you did but the token was invalid -

```
❯ ./influx user ls
Error: failed to list users: 401 Unauthorized: unauthorized access
```

This is confusing, especially when you think you have specified the token but accidentally misspelt the env var.

This PR causes the CLI to return a specific error instead -

```
❯ ./influx user ls
Error: influx token required
```

I also added some tests; not sure if they're needed/wanted because there aren't any existing tests in the package yet. I'm happy to drop them.